refactor: properly name `autocomplete` dataFilter
